+++ lib/steps/step_06_move_files/moving_strategies/reverse_shortcut_moving_strategy.dart
@@ -200,6 +200,7 @@ class ReverseShortcutMovingStrategy extends MoveMediaEntityStrategy {
                   allPhotosDir,
                   anchorMoved,
                   desiredName,
+                  context.hardlink,
                 );
             ssw.stop();