+++ lib/steps/step_06_move_files/services/step_06_move_media_entity_service.dart
@@ -515,6 +515,7 @@ class MoveMediaEntityService with LoggerMixin {
       albumBehavior: context.config.albumBehavior,
       allPhotosDirectoryName: context.config.allPhotosDirectoryName,
       dividePartnerShared: context.config.dividePartnerShared,
+      hardlink: context.config.hardlink,
     );
 
     int entitiesProcessed = 0;